+++ packages/ub/CHANGELOG.md
@@ -17,9 +17,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   req.parsedParameters // { foo: 'bar', baz: ['qux', 'quux'], corge: '' }
   ```
   We recommend using this getter instead of `querystring.parse(req.parameters)` to prevent multiple
-  call to parameter parsing from different methods.
+  call to parameter parsing from different methods (require **UB server >= 5.19.0**).
 
 ### Changed
+ - **BREAKING** JS endpoints, added by `App.registerEndpoint` and native endpoints, added by a server (stat, auth, ubql, logout and metrics)
+  now executed using `App.launchEndpoint` JS implementation.
+   
+  This allows to use the same `req` and `resp` objects for both endpoint types, and a one step forward to pure JS ubql implementation.
+
+  These changes require **UB server >= 5.19.0**.
 
 ### Deprecated