+++ packages/ub/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.19.1] - 2021-02-03
 ## [5.19.0] - 2021-02-02
 ### Added
  - `THTTPRequest` extended by helper functions: